+++ release-notes.md
@@ -5,6 +5,7 @@
 - Add support for tearing.
 - Add support for touch input.
 - Add support for libei.
+- Add support for RemoteDesktop portal.
 
 # 1.4.0 (2024-07-07)